server/pyproject.toml
[tool.poetry]
name = "whai"
version = "0.1.0"
description = ""
authors = ["plutokekz <Lukas.Mahr@campus.lmu.de >"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.95.1"
pydantic = "^1.10.7"
uvicorn = "^0.22.0"
fpdf = "^1.7.2"
pypdf = "^3.8.1"
aspose-words = "^23.4.0"
pdfkit = "^1.0.0"
azure-cognitiveservices-speech = "^1.27.0"
python-dotenv = "^1.0.0"
requests = "^2.29.0"
azure-ai-textanalytics = "^5.2.1"
msrest = "^0.7.1"
httpx = "^0.24.0"
python-multipart = "^0.0.6"
pydub = "^0.25.1"
cohere = "^4.3.0"
openai = "^0.27.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
